requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
blas=1.0=mkl
bottleneck=1.4.2=py311h57dcf0c_0
brotli-python=1.0.9=py311h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
contourpy=1.3.1=py311h214f63a_0
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.55.3=py311h827c3e9_0
freetype=2.12.1=ha860e81_0
icc_rt=2022.1.0=h6049295_2
intel-openmp=2023.1.0=h59b6b97_46320
joblib=1.4.2=py311haa95532_0
jpeg=9e=h827c3e9_3
kiwisolver=1.4.8=py311h5da7b33_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=h44ae7cf_1
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
matplotlib-base=3.10.0=py311he19b0ae_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h827c3e9_2
mkl_fft=1.3.11=py311h827c3e9_0
mkl_random=1.2.8=py311hea22821_0
numexpr=2.10.1=py311h4cd664f_0
numpy=2.0.1=py311hdab7c0b_1
numpy-base=2.0.1=py311hd01c5d8_1
openjpeg=2.5.2=hae555c5_0
openssl=3.0.16=h3f729d1_0
packaging=24.2=py311haa95532_0
pandas=2.2.3=py311h5da7b33_0
pillow=11.1.0=py311h096bfcc_0
pip=25.0=py311haa95532_0
pyparsing=3.2.0=py311haa95532_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0post0=py311haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py311haa95532_0
scikit-learn=1.6.1=py311h585ebfc_0
scipy=1.15.1=py311h9f229c6_0
seaborn=0.13.2=py311haa95532_2
setuptools=75.8.0=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.5.0=py311h746a85d_0
tk=8.6.14=h0416ee5_0
tzdata=2025a=h04d1e81_0
unicodedata2=15.1.0=py311h827c3e9_1
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wheel=0.45.1=py311haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
